fix(utils): Rename global.ts -> globalobject.ts
