+++ packages/utils/src/index.ts
@@ -1,7 +1,7 @@
 export * from './browser';
 export * from './dsn';
 export * from './error';
-export * from './global';
+export * from './globalobject';
 export * from './instrument';
 export * from './is';
 export * from './logger';
