+++ packages/utils/src/logger.ts
@@ -1,6 +1,6 @@
 import { WrappedFunction } from '@sentry/types';
 
-import { getGlobalSingleton, GLOBAL_OBJ } from './global';
+import { getGlobalSingleton, GLOBAL_OBJ } from './globalobject';
 
 /** Prefix for logging strings */
 const PREFIX = 'Sentry Logger ';
